Fix the terrain shaders on macOS
Both terrain shaders declared `flat varying uint TexIndex`, which Apple's
GLSL 1.20 compiler rejects with a syntax error even though it advertises
GL_EXT_gpu_shader4 and accepts usampler2D, texelFetch2D and
gl_InstanceIDARB from that same extension. Probing each construct
separately showed the uint varying is the only one it refuses.

TexIndex is a float now -- still flat, so it is not interpolated, and a
float carries tile indices exactly. Terrain had never rendered on macOS:
neither shader compiled, so the ground outside a dungeon was black.

+++ data/shaders/terrain.frag
@@ -12,11 +12,12 @@ varying vec3 n_viewspace;
 varying vec3 t_viewspace;
 varying vec3 b_viewspace;
 varying vec4 TexCoords;
-flat varying uint TexIndex;
+// See terrain.vert: a uint varying does not compile on macOS.
+flat varying float TexIndex;
 
 void main()
 {
-    vec3 coord = vec3(TexCoords.xy, float(TexIndex));
+    vec3 coord = vec3(TexCoords.xy, TexIndex);
     vec4 color = vec4(texture2DArray(diffuseTex, coord).rgb, 0.0);
     vec4 nn = vec4(0.5, 0.5, 1.0, 1.0);
